Consolidate datetime functions
