--- src/utils/dateTime.ts
@@ -1,2 +1,0 @@
-export const isoDate = (isoDateTime: string): string =>
-  isoDateTime.slice(0, 10);
